# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"Biblioteca","description":"Documentação da Biblioteca API","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Autor","description":"Gerenciador de autores"},{"name":"Empréstimo","description":"Gerenciador de empréstimos"},{"name":"Empréstimo associados","description":"Visualizador de empréstimos associados"},{"name":"Livro","description":"Gerenciador de livros"},{"name":"Cliente","description":"Gerenciador de clientes"}],"paths":{"/v1/loan":{"get":{"tags":["Empréstimo"],"summary":"Listar todos os empréstimos","description":"Lista todos os empréstimos cadastrados com as associações","operationId":"listAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Loan"}}}}}}},"post":{"tags":["Empréstimo"],"summary":"Registrar um novo empréstimo","description":"Registra um novo empréstimo a partir do preenchimento de um DTO","operationId":"registerLoan","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLoanDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Loan"}}}}}}},"/v1/loan/{id}/client/{clientId}":{"post":{"tags":["Empréstimo"],"summary":"Associar empréstimo com cliente","description":"Associa um empréstimo já registrado com um cliente já cadastrado","operationId":"associateLoanClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"clientId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Loan"}}}}}}},"/v1/loan/{id}/book/{bookId}":{"post":{"tags":["Empréstimo"],"summary":"Associar empréstimo com livro","description":"Associa um empréstimo já registrado com um livro já cadastrado","operationId":"associateLoanBook","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"bookId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Loan"}}}}}}},"/v1/client":{"get":{"tags":["Cliente"],"summary":"Listar todos os clientes","description":"Lista todos os clientes cadastrados no banco de dados","operationId":"listAll_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClientResponseDto"}}}}}}},"post":{"tags":["Cliente"],"summary":"Registrar um novo cliente","description":"Registra um novo cliente a partir do preenchimento de um DTO","operationId":"registerClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClientDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Client"}}}}}}},"/v1/book":{"get":{"tags":["Livro"],"summary":"Listar todos os livros","description":"Lista todos os livros cadastrados no banco de dados","operationId":"listBooks","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Book"}}}}}}},"post":{"tags":["Livro"],"summary":"Registrar um novo livro","description":"Registra um novo livro a partir do preenchimento de um DTO","operationId":"registerBook","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Book"}}}}}}},"/v1/book/{id}/author/{authorId}":{"post":{"tags":["Livro"],"summary":"Associar livro com autor","description":"Associa um livro já registrado com um autor já cadastrado","operationId":"associateBookAuthor","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"authorId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Book"}}}}}}},"/v1/author":{"get":{"tags":["Autor"],"summary":"Listar todos os autorees","description":"Lista todos os autores cadastrados no banco de dados","operationId":"listAuthors","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Author"}}}}}}},"post":{"tags":["Autor"],"summary":"Registrar um novo autor","description":"Registra um novo autor a partir do preenchimento de um DTO","operationId":"registerAuthor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAuthorDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Author"}}}}}}},"/v1/loanBook":{"get":{"tags":["Empréstimo associados"],"summary":"Listar os empréstimos associados","description":"Lista todos os empréstimos cadastrados com as associações","operationId":"listAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LoanBookResponseDto"}}}}}}}},"/v1/client/{id}":{"delete":{"tags":["Cliente"],"summary":"Deletar um cliente","description":"Deleta um cliente do banco de dados a partir do id dele","operationId":"deletedById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/v1/book/{id}":{"delete":{"tags":["Livro"],"summary":"Deletar um livro","description":"Deleta um livro do banco de dados a partir do id dele","operationId":"deletedById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/v1/author/{id}":{"delete":{"tags":["Autor"],"summary":"Deletar um autor","description":"Deleta um autor do banco de dados a partir do id dele","operationId":"deletedById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"CreateLoanDto":{"type":"object","properties":{"description":{"type":"string"}}},"Client":{"type":"object","properties":{"clienteId":{"type":"string","format":"uuid"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"integer","format":"int32"}}},"Loan":{"type":"object","properties":{"loanId":{"type":"string","format":"uuid"},"loanDate":{"type":"string","format":"date-time"},"description":{"type":"string"},"client":{"$ref":"#/components/schemas/Client"}}},"CreateClientDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"integer","format":"int32"}}},"CreateBookDto":{"type":"object","properties":{"name":{"type":"string"},"genre":{"type":"string"},"year":{"type":"integer","format":"int32"}}},"Author":{"type":"object","properties":{"authorId":{"type":"string","format":"uuid"},"name":{"type":"string"},"nationality":{"type":"string"}}},"Book":{"type":"object","properties":{"bookId":{"type":"string","format":"uuid"},"name":{"type":"string"},"genre":{"type":"string"},"year":{"type":"integer","format":"int32"},"authors":{"type":"array","items":{"$ref":"#/components/schemas/Author"}}}},"CreateAuthorDto":{"type":"object","properties":{"name":{"type":"string"},"nationality":{"type":"string"}}},"LoanBookId":{"type":"object","properties":{"loanId":{"type":"string","format":"uuid"},"bookId":{"type":"string","format":"uuid"}}},"LoanBookResponseDto":{"type":"object","properties":{"clientId":{"type":"string","format":"uuid"},"loanBookId":{"$ref":"#/components/schemas/LoanBookId"},"clientName":{"type":"string"},"email":{"type":"string"},"bookName":{"type":"string"},"authors":{"type":"array","items":{"type":"string"}},"loanDate":{"type":"string","format":"date-time"},"description":{"type":"string"}}},"ClientResponseDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"}}}}}}